+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Refactored `EuiSuggest` to use `EuiSelectable` ([#5157](https://github.com/elastic/eui/pull/5157))
 - Added a return type to `EuiTable` `resolveWidthAsStyle` util ([#5615](https://github.com/elastic/eui/pull/5615))
 - Added a return type to `euiSelectableTemplateSitewideFormatOptions` util ([#5620](https://github.com/elastic/eui/pull/5620))
+- Added `isDisabled` prop to `EuiMarkdownEditor` ([#0000](https://github.com/elastic/eui/pull/0000))
 
 **Bug fixes**